eazeaea

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-int suM;
+unsigned long suM;
 int k;
 
 for (k = 2; k < 50; k++)
